Fix minimal Xcode toolchain select
Summary:
# Problem

There's a constraint that is used to signal selection of the Minimal Xcode toolchain. But that constraint is not more specific than the toolchain select constraints, so it leads to a problem of resolving a platform `select()` because of ambiguity (see *Test Plan*).

# Solution

We implement prioritisation using nested `select()`s. Specifically, we prefer Minimal Xcode over a specifically selected toolchain.

Reviewed By: IanChilds

Differential Revision: D36443090

fbshipit-source-id: c052508cffc5ec1d1d6544e6c8ec2d075cf89a05