Do not enable local Linux execution on Windows
Summary:
The logic whether to enable local execution on Linux exec platforms was based on the expression `not host_is_mac` which would produce the wrong value when Buck2 runs on Windows.

Instead, we want to enable local Linux execution when running on a Linux machine.

The original code was added before we had a Windows exec platform, so it made sense at the time.

Reviewed By: IanChilds

Differential Revision: D36510483

fbshipit-source-id: 683dce6bccf157cb3c95216d7dd32a0550e1d55b