Support `java_for_tests` toolchain param
Summary:
v1 supports a `java_for_tests` toolchain parameter, which provides the java
tool to use when running tests, in cases where this is different from the one
used for non-tests.  This diff adds support for it in v2.

Tests in fbcode use this to use a shim which redirects the JVM process to the
fbcode platform runtime:

```
$ buck2 test mode/devo-nosan //infrasec/authorization/java/aclcheckerhandler_jni:ThriftAclCheckerHandlerJNITests-java11
```

Reviewed By: IanChilds

Differential Revision: D36467830

fbshipit-source-id: 83a1547b3fb65337690dd26b7de712feaaf69c36