Get targets to build from target attribute data
Summary:
Generated source files (files in buck-out) have a target from which they're built. When you perform a query (the backing data for TargetAttributes), we have the target along with any subtarget on that target. However, data deriving from `xcode-data` does not have the subtarget. This is due to a limitation in buck2, in which we only know the deriving target of an artifact, not the deriving target and any subtarget/named output that might be defined.

This meant that in the TargetMerger, when we sought to add targets to prebuild, we were only ever able to add the target, not the target + subtarget.

In some of the Messenger genrules, the default output is the header file, but to obtain the implementation file, you must request it by subtarget/named output.

In the case of the Messenger workspace, this resulted in us not prebuilding a large number of implementation files, and they would appear disabled and missing in Xcode.

This is not a problem for non-merged targets, which were getting the targets to build from the target attribute data.

To fix this problem in the short term, we pass the target attribute data down and through some finagling, find the actual target+subtarget to build. Longer term, we should structure this data better and split out how/when we determine which targets to build.

Reviewed By: milend

Differential Revision: D36522919

fbshipit-source-id: e50bfd8acae6b449e75ae2f04c1b2fe3732d50fb