Add answer-checking functionality to fuzzer
Summary:
This is complicated slightly by the fact that sometimes panicking is the correct behavior for DICE - say we set InjectedKey1 at v1, set InjectedKey2 at v2, and then try to read InjectedKey2 at v1: that should panic, so we need to check that and catch it.

The implementation of this is naive, since it doesn't re-use shared values computed at the same version. In practice I think this is fine because most fuzz test cases are computing values at wildly different versions, and also are qutie small.

Reviewed By: bobyangyf

Differential Revision: D36359914

fbshipit-source-id: 26e36b6157f35706951aea85e467f1406141c195