Dump node pointers to disambiguate nodes at different versions
Summary: Previously, we only logged a KeyID for each node, and the deps and rdeps as a pair of (KeyID, version). This was fine for deps - we do, in fact, internally store deps as a pair of (Key, Version) - but we store rdeps as pointers. Right now, in theory, the LastN storage config we use means that there'll be only one live node with a key that depends on the rdep. But if there were, hypothetically, multiple nodes, our DICE dump wouldn't tell us which node-with-a-key a given rdep pointed to.

Reviewed By: bobyangyf

Differential Revision: D36456958

fbshipit-source-id: 40c93322d9a9987123fd8428d9cf92c941960cc8