buck2: allow access to actual stdout / stderr
Summary:
Right now we mess with stdout and stderr a lot:

- We do a lossy conversion to String
- We truncate
- We swallow errors

This is all fine when users are the recipients of this output, but when
machines are (which they are if we run tests on RE because some tests give
their test listing on stdout), then it doesn't work as well.

This diff fixes this by replacing `stdout: String` and `stderr: String` with
structured data that we can access on demand to retrieve what we actually need.

A few notes:

- The action execution stuff had to change a bit to accommodate the fact that
  we can't just `take()` stderr out of metadata anymore. I refactored this to
  simplify the big match we had.
- This is a big diff but we can't easily change the output without changing the
  places it's consumed.
- We do this thing right now where we always fetch stderr and prefetch it out of RE
  on the success path if needed. I implemented this in the new code with a dedicated
  variant for "this was a digest and we prefetched this String out of it". I didn't optimize
  the codepath where we download the output here then later want to fetch the full stderr.
- There is a behavior change: if an action fails to produce all outputs,
  we now show stdout in addition to stderr, which is neat and helpful and
  correct, though it means we have to wait for stdout.
- The end goal of this is that tests now access the actual output, not a
  potentially-munged version thereof.

TL;DR: current code paths are unchanged, new codepath allows for accessing stdout
and stderr in an unadulterated format.

Reviewed By: ndmitchell

Differential Revision: D36347817

fbshipit-source-id: a57006ff54744495ed71862dd927bdb4a0da95fe