Save memory by using write_json to encode target_label.
Summary:
NOTE: This requires D36506684 (https://github.com/facebookincubator/buck2/commit/d0ab00380dd9a9e738f6b6168137f71eaea76b05) to be shipped in the buck2 executable, I'll land this tomorrow.

In the latest heap profile of FBiOS, we can see that `str(label.raw_target())` was accoutning for 1.7 GB of the 37 GB. With nmitchell's latest change - we don't have to create a string. Instead, we can pass `target_label` to the `write_json` API and have it do the encoding for us.

Reviewed By: chatura-atapattu

Differential Revision: D36521093

fbshipit-source-id: d06116380f998cd825f3d6eb584a1f43c446cc1c