Fixing js_bundle operations in buck2.
Summary:
When migrating `js_library` + `js_bundle` to use `write_json`, I failed to notice that `extraData` should be encoded as a JSON object, not a string.

Unfortunately, there is no mechanism to convert a JSON-string to a JSON-object in buck2 (yet), so, lets add a well-documented "HACK" to fixup the JSON so that all the scripts are happy.

T121096376 will track this issue long-term. When/if we ever support it, we can remove all this stuff.

## why don't you patch the scripts? is this necessary?
I tried that, but all of the bundling tools which require the patch are open-source via `facebook/metro`.

I followed-up after the [comment from moti](https://www.internalfb.com/diff/D36482836?dst_version_fbid=296022886076087&transaction_fbid=1026790477960673), and realized there is nowhere to do this between `write_json` and the open-source code.

Reviewed By: IanChilds

Differential Revision: D36482836

fbshipit-source-id: 0357dcd826f11a4121210b66973c50b4b0745e09