Use experimental_regex instead of regex_match in class_name_filter
Summary:
This is more performant, so move to it. I did some benchmarking on how long this filtering would take on Messenger (see D36547354), and previously it was 271s, and now it is ~18s. That's still a bit too slow for what we're going to need it for (this is going to run on nearly every build), so we might still need a solution like D35432374. For now, this will do.

I tried to profile to see where all the time is going, I was running (with D36547354 in place):
```
buck2 kill && BUCK2_STARLARK_PROFILE=fbcode//buck2/prelude/android:class_name_filter.bzl=FLAME buck2 build fbandroid/buck2/tests/good/dex:
```
but I didn't seem to get any output file in `fbsource`, so not sure what I was doing wrong there?

Reviewed By: ndmitchell

Differential Revision: D36547305

fbshipit-source-id: 3cd72e3fe83606f4767ff326908155dcd5c0d1ce