Add wall_time to ActionExecutionEnd
Summary:
This takes the wall_time that we previously used for critical-path calculations, and includes it with every action execution. This is different from the SpanEnd's `duration` because it omits queueing and anything else that isn't the action.

This is, in some sense, a bit redundant: the CLI already has access to this data because execution stages are reported, but there are two reasons we can't use this in `buck2_actions`:
1. We don't send execution stages to Scribe, even just sending execution end events would 1.5x our traffic instantly for little gain
2. Even if we did (1.), since the execution-end span arrives at the tailer before the action-end does, and we can only write a row to Scuba when the action-end event arrives, the tailer would need to hold execution durations in memory, keyed by action, until completion, which would increase memory pressure + add state that makes it harder to recover from a crash-cycle without data loss.

Reviewed By: krallin

Differential Revision: D36496102

fbshipit-source-id: c7a83ccb84069c227f67a1b8b7271099f2956b9f