EqConst and friends
Summary:
Add four more instructions:
* EqConst — generic `a == constant`
* EqPtr — `a.ptr_eq(constant)`
* EqStr — `a == constant` where constant is a string
* EqInt — `a == constant` where constant is `i32`

`Eq` is frequent instruction (about 6% of all instruction), and it is often evaluated with one of the arguments constant. So it is helpful to have specialized instructions for equality.

Note generic `a.equals(b)` is quite expensive. E. g. when evaluating `a == None` (e.g. in `x.get("k") != None`) where `a` is a string, we perform:
* identity comparison first (false)
* `x.get_ref` with branches for `i32` and non-`i32`
* virtual call `x.get_ref().equals(...)`
* `y.is_str` inside `StarlarkStr::equals`
while it could be (and it is after this diff) just `a.ptr_eq(b)`

This diff introduces a bug, where strings of length <= 1 are considered unique, but they may be not unique, this is addressed in the following diffs.

Reviewed By: ndmitchell

Differential Revision: D36580816

fbshipit-source-id: 833fe8c53d1552235b391167d894ff22b43be267