Make rustdoc compression runnable in remote execution
Summary:
For the purpose of the intern rustdoc site, making `…[doc]` subtargets run the real rustdoc is not ideal because it results in an enormous number of tiny html files from Remote Execution needing to be materialized onto the local machine for compression. Buck/eden can only write so many files per second into buck-out, and this is by far the current bottleneck, accounting for 70% of the runtime of the daily docs rebuild job.

This diff adds a target that wraps the rustdoc generation and compression together into a single step, emitting only the compressed archive and search index from Remote Execution.

This target can be dropped in as a replacement for the rustdoc-wrapper normally run by `doc` subtargets, by doing `buck2 build --config rust.rustdoc=fbcode//third-party-buck/platform010/build/rust:rustdoc-compressed`.

As a side benefit, all of the compression now happens in parallel in Remote Execution instead of serially on the small number of cores available in a single Sandcastle host, but that is not (currently) a significant part of the runtime.

Reviewed By: zertosh

Differential Revision: D36584815

fbshipit-source-id: 0039d92f9b18bf27d1c0d63d356dc5aa35769200