bucK2: rename ctx.label.test_cwd -> cell_root (keep it for backwards compat)
Summary:
A week or so ago I added `relative_to(ctx.label.test_cwd)`. In hindsight, I
think this should be `ctx.label.cell_root`. The reason is that the logical
place the define where a test runs from would be its test info, but that isn't
available when we look at `relative_to`!

This means test authors would have to keep what they put on their test info and
what they make their artifacts relative to in sync, but I think ultimately it's
the better approach.

This will let us have the cwd be a parameter on ExternalRunnerTestInfo, which
is much nicer since it means we can e.g. use this to decide where a given test
should run from. It's also a lot less implicit.

Those are nice to haves, but this will also help with some work TAE wants to do
around testing test frameworks where this is actually blocker. Indeed, idea
there is to capture and serialize the test info then re-export a new test (so
the test-test is a UDR that wraps an existing test).

For that to work properly, the re-exported test will have to use the same cwd
as the original test, but if we control this via labels (the original plan),
then the re-exported test doesn't control those. If we do it via
ExternalRunnerTestInfo instead, then it can!

Reviewed By: ndmitchell

Differential Revision: D36545192

fbshipit-source-id: a109ea98d90b58e8544c9df0b81b07def9102ee7