buck2: only prefetch stderr when running actions
Summary:
Like it says in the title.

When running tests, we want the whole thing, raw, and we don't have a deep
dependency tree (at most we are doing listing and there is one layer of test
runing to happen later), so micro-optimizing latency isn't particularly
important.

Besides, long-term, we don't even want to download output for tests (and
instead we'll just give the digests to tpx).

Reviewed By: bobyangyf

Differential Revision: D36526250

fbshipit-source-id: 57c65e7dec0be6606a0cd9824ee8ed403598a588