Adjust Linux mode files
Summary:
We don't want to be using `--fake-host` for building, so let's change `fbsource//fbobjc/mode/buck2/linux` so that it actually adjusts the available platforms instead, as that's all which is needed.

Keep the old mode file around with a new name `linux-fake` for the time being while transitioning things over (still useful for querying).

Reviewed By: blackm00n

Differential Revision: D36591902

fbshipit-source-id: 6691086755532b45c566f2fd076de8171d85ac0f