buck2: test: add --unstable-force-tests-on-re
Summary:
Like it says in the title. This will force tests to run from the project root
and with relative paths, which is the only mode we currently support on RE.

This is basically the behavior I'd added in `--unstable-allow-tests-on-re`
before I made running from the project root and with relative paths options on
ExternalRunnerTestInfo. For the time being, this is useful to play around
with tests on RE without having to actually reconfigure anything.

Reviewed By: ndmitchell

Differential Revision: D36592134

fbshipit-source-id: d79ae9c0707a3d19e2a8b1a19dd24fd88c765f00