buck2: implement use_project_relative_paths, run_from_project_root
Summary:
Like it says in the title. Make those flags actually do something.

Note that if we have use_project_relative_paths = True, run_from_project_root =
False, we produce paths that are probably wrong. This is mostly intentional,
since rule authors _can_ fix it using `relative_to(ctx.label.cell_root)`, but
also they probably shouldn't do this as it's a clowny combo.

This is, mainly, why we don't need *only* `use_project_relative_paths` for this
to work.

Note: this breaks `--unstable-allow-tests-on-re` since now most tests won't
have the right paths when running on RE. This is fixed in the next diff:

- I'll make that flag only allow RE if those conditions are met.
- I'll add a new flag, `--unstable-force-tests-on-re` which defaults those two fields to true AND enables `--unstable-allow-tests-on-re`, thus forcing tests to attempt to run on RE.

Reviewed By: ndmitchell

Differential Revision: D36545911

fbshipit-source-id: b33adbf6781cc27d67052ac56529e9e622ec8e48