Do jar compression without having to call MultiDexExecutableMain
Summary:
Right now, if we are doing jar compression on our secondary dexes, then we create them as follows:
1, Create the raw dexes (in parallel), e.g. `classes.dex`, `classes2.dex`, `classes3.dex`.
2. Pass all the raw dexes to `MultiDexExecutableMain`, which will create jars for each of them in turn, add a `.meta` file for each and write `metadata.txt`, and put all of that into a "secondary dex dir".

This is quite inefficient - this step takes > 20 seconds in my local runs when building Messenger. Since we'll have to do this whenever any Java code changes, that's really expensive.

Instead, in step (1) we create the secondary dexes directly, along with the `.meta` file and the line of metadata info that needs to be added to `metadata.txt`. We then have a second dynamic action that reads all the lines of metadata info and write them to `metadata.txt`.

This way, we remove the need for `MultiDexExecutableMain` entirely. This should result in a reasonably significant perf speedup for incremental builds.

Reviewed By: cjhopman

Differential Revision: D36476795

fbshipit-source-id: e8af7bc47667bbcd0d1a779df7b1579e2f5b85a9