buck2: be consistent when ordering log files
Summary:
In D36483960 (https://github.com/facebookincubator/buck2/commit/12fb5852a5fb848ac561016204fd09725fdedf04) we renamed our log files and stopped caring about filenames and
used creation times instead to sort logfiles, but as it turns out we have this
logic duplicated so we started looking at the wrong files sometimes :|

The names still have timestamps in them so this happens to work out most of the
time, but in tests it does break them.

This diff deduplicates the logic to fix the bug.

Reviewed By: ndmitchell

Differential Revision: D36629981

fbshipit-source-id: dd52de0fad71eeea2658d70845984e44ecd610ba