Make all but first two rule function parameters named-only
Summary:
All parameters except `implementation` and `attrs`. These parameters are optional, so it's important to not allow them to be positional-only.

`implementation` and `attrs` are positional-or-named after this diff (not changed).

`attr` is always passed by name, so maybe it should be named-only too.

`implementation` is also always passed by name, but maybe it should be positional-only.

Reviewed By: krallin

Differential Revision: D36616108

fbshipit-source-id: 57b7abac89726775871af88c8330b17bf8d7c072