pass -opaque to compile steps in dev modes
Summary: [the ocaml manual](https://v2.ocaml.org/manual/native.html) explains that the compile flag `-opaque` will degrade runtime performance (for lack of cross module inlining) but can lead to faster edit/build cycles so this diff arranges to pass this flag for dev modes. when compiling this way, we make it so that '.cmx' files depend only on '.cmi' files and not other '.cmx' files wherever possible.

Reviewed By: ndmitchell

Differential Revision: D36566211

fbshipit-source-id: 9c73e1aa89a9ca0444a43f656e45d1f40092862b