When updating cache, check for nearest entry rather than just the previous
Summary:
If we only check the previous keys, if requests to dice arrive out of order, we will fail to properly deduplicate and obey the "lastn" storage policy because the versioned maps would never check the newer versions of this key.

To fix this, we check both forward and backwards, picking the nearest version.

Reviewed By: jaspercb

Differential Revision: D36117243

fbshipit-source-id: 750bb798e60a7dce97685b97e905795ccc8a5ab7