Make EventDispatcher available implicitly
Summary:
This commit introduces a mutable global thread-local events dispatcher and an accompanying public API to implicitly use the global dispatcher downstream for `span_async`, `span`, and `instant_event` after being set.

The added tests exercise the following:
1. Multiple concurrent event dispatchers (separately used by 2 threads)
2. Nested spans.
3. Context-switching between threads and verifying that the global dispatcher is the one initially created by the thread.
4. Verifying that all events are sent, and to the correct dispatcher.

The above is tested with both single-threaded and multi-threaded tokio runtimes.

The diff is ~100 lines of implementation, and the rest is of tests (refactoring + adding 2 tests).

Reviewed By: swgillespie

Differential Revision: D36586452

fbshipit-source-id: fd35e09304b7d40d1b955559f28500ab8812a953