buck2: split up `FileDigest` into `FileDigestData` and a TTL tracker
Summary:
In this stack, I'm introducing a base digest type (this is eventually called `FileDigest` but first
it becomes called `FileDigestData`), and a digest + TLL type (this is currently `FileDigest` and
later it becomes `TrackedFileDigest`).

it makes sense to serialize both of those to RE types, but it doesn't make sense to parse
digests directly into what will become `TrackedFileDigest`: we should only allocate TTL
tracker for digests that we intend to access for some time.

So, this diff introduces this new "not-tracked" type (`FileDigestData`) and converts places
that just need to parse digests to use that instead. It also splits our Buck2 <> RE conversion
trait to no longer allow going directly from a RE digest to the tracked thing.

Again, the names are not what they are at the end of this stack but doing the extraction
before the rename is easier than doing it at the same time :)

Reviewed By: ndmitchell

Differential Revision: D36511776

fbshipit-source-id: dbb1762d64bb56c92313a1d8f04aa8ea5328e88a