buck2: TrackedFileDigest: implement `Borrow<FileDigest>` and use it in sets
Summary:
This means you don't need an expiry tracker to access a set of
TrackedFileDigest, which is quite useful.

For the sake of safety, I manually implemented the various things we proxy to
`data()` since ignoring them was getting verbose and a bit error-prone. I wish
Derivative had a "use that field for this derive" but it doesn't.

Reviewed By: bobyangyf

Differential Revision: D36511768

fbshipit-source-id: 7c7e26d513a5ddee99045431a52e02b450c66091