buck2: re stub: add a stub for get_digests_ttl
Summary:
This stack
===

The end goal of this stack is to track the TTL of all FileDigests known to Buck2.

D36511777 explains why that is a good thing and shows how that impacts performance.

This diff
===

Like it says in the title. I'd like to use this RE method so let's make it available.

Reviewed By: ndmitchell

Differential Revision: D36511766

fbshipit-source-id: b57fabd679fed025c4004b6be713ccce1db08de3