buck2: respect the TTLs returned by RE on execute & action cache hits
Summary:
Follow up to D36511763. Rather than set a fixed 5 hour TTL for everything we retrieve
from RE, make the TTL whatever RE says it is.

That said, is the diff that puts it all together so I'll put an overview of perf here as well.
Note that we already have some client-side tracking of digests in the form of a cache
in the RE client, so what we're expecting here is:

-  improved CPU utilization (because we  just need a pointer lookup to read TTLs
   rather than a shared . hashmap lookup)
- Possibly slightly better performance (i.e. reduction in lookups) since we can tell digests we
  still care about from digests we don't care about (whereas a LRU cache can't), and since
  we populate TTLs when downloading action results (so for those, we always eliminate 1
  lookup, which the caching approach has to do).

Reviewed By: bobyangyf

Differential Revision: D36511777

fbshipit-source-id: 68a9540c2df79d606503a51226a9bdee40359405