buck2: track and use last known expiration timestamp on FileDigest
Summary:
Note: there is more refactoring to make this more palatable later in this stack
(in fact, it's most of this stack).

This updates FileDigest to include a field to track the last known TTL of a digest, and
to skip asking RE about digests that are known to expire a long time from now.

Note that the implementation is somewhat naïve here and does not deduplicate
concurrent calls. Perf testing on D36511777 shows that this still yields a (very) sizeable
reduction in digests we request from RE.

Reviewed By: ndmitchell

Differential Revision: D36511767

fbshipit-source-id: fbbab6768f127c44a6558aab5d0f5547d784959e