buck2: use the RemoteExecutorUseCase from RemoteExecutorOptions
Summary:
Like it says in the title. This uses the RemoteExecutorUseCase that was passed
via RemoteExecutorOptions when making calls to RE.

This isn't used *everywhere* so sometimes we do fallback to the global default,
but it is used specifically on the path of running an action (uploading,
executing, and downloading outputs), which is where it matters most.

Reviewed By: bobyangyf

Differential Revision: D36627059

fbshipit-source-id: d6a945fbc73ba229c7a30899c954133fe88321db