buck2: rename CommandExecutorConfig -> CommandExecutorRequest
Summary:
Like it says in the title.

I'd like to rename `ActionExecutorConfig` to `CommandExecutorConfig`, so I need
to free up the name first. The reason I want to do this rename is because this
config only controls *command* execution (commands being things we run, as
opposed to actions that might be e.g. write a file or create a symlink dir).

Since `ActionExecutorConfig` is used quite a bit, and `CommandExecutorConfig`
is barely used at all, let's free the name for the more important thing.

I renamed this to `CommandExecutorRequest` since it's used to request a
command executor.

 ---

The end goal of this refactoring is to allow: D36627058

Reviewed By: bobyangyf

Differential Revision: D36627065

fbshipit-source-id: 99b3ab19c0e813624b3acdb9d7065d017a405c45