Add more tests for full classpath
Summary:
Adds new tests for classpath, we now tests for targets, outputs and direct usage of `$(classpath)`. For targets, we want all targets regardless of whether the target has an output, for the other two we just get the targets that have an output.

Moved to `TARGETS` since we can run this with `buck1`. Also commented out the failing test, will be fixed later in the stack.

Reviewed By: cjhopman

Differential Revision: D36630074

fbshipit-source-id: 4a9cb8bedbe6f8debb15d84e1ecea6dafab0412d