try and connect using uds if that fails fall back to tcp
Summary:
For installs on macos, we need to fallback to using tcp because GRPC java code does not currently support uds. It is possible, if kQueue is the transport selected to use uds on macos. However, the grpc-io-nettyshaded jar used in buckv1 does not have these classes. It does have access for Epoll which enables the use of uds for linux installs.

There was also [discussion](https://fb.workplace.com/groups/buck2dev/permalink/3167410723546978/) on how we should have a more generic fallback structure for all connections. So the buckd will also fallback to using tcp, if uds fails.

Reviewed By: cjhopman

Differential Revision: D36083330

fbshipit-source-id: 3a243ef6fbaedd1afa39095cfe9fdb62430aa910