classpath command should use classpath_including_targets_with_no_output template
Summary: The `$(classpath)` macro has to return only those targets with an output (e.g. in case people actually want to use it as a Java classpath in a genrule). For the `classpath` function we need to also include targets that don't have a classpath, so use this template info instead.

Reviewed By: cjhopman

Differential Revision: D36630117

fbshipit-source-id: 3c6b5d3252a462f3873f8a9bd0278003e005a912