Add SplitDwarfInfo to cxx_toolchain
Summary:
`apple_(binary|library)` shares a codepath with `cxx_executable`, `cxx_link` and more. Within those codepaths there is hardcoded support for split DWARF, which is not supported for darwin builds. Our debug info goes in a DSYM.

To surgically fix this, lets expose `split_dwarf_info` on the toolchain and read whether it is configured there. That will enable us to force-disable it for Apple's `cxx_toolchain`.

Reviewed By: chatura-atapattu

Differential Revision: D36669130

fbshipit-source-id: b9243a6bc822de6854d06c40d32a622c8b57ce63