classpath function output should include libs with no output
Summary:
Have mixed feelings about this, but this is the best way I can think to solve this. Happy to hear other ideas!

Problem (see T120950937): suppose `A` depends upon `B`, and `B` exports `C`, but `B` has no output. In `buck1`, `classpath(A)` would give `[A, B, C]`, but in `buck2` it will give `[A, C]`. This means that if you try to do e.g. `deps_query = "classpath(A) intersect B"`, you'll get a different answer in `buck2` (and compilation will fail if you depend upon something in `C`, since it won't be a dep in `buck2` but it will in `buck1`.

This fix makes it so that `JavaPackagingDep` is augmented with an extra field `output_for_classpath_macro` (which is used to map from artifact -> target in the core). We then create a `JavaPackagingDep` for every library, even if it has no output, and the set of all packaging deps is the set that the `classpath` function should return. The bit that is ugly is that now we have `JavaPackagingDep`s for things that aren't actually packaged.

Reviewed By: cjhopman

Differential Revision: D36515349

fbshipit-source-id: 8fbf751c464382688b53362f6dddd8072123cbfd