Change the way how the zippy client is configured
Summary:
Using the zippy client mode instead of the deprecated boolean flag. Modes allow to use the hybrid client.

Hybrid Client:
We are rolling out the zdb hybrid client which tries to use as much as possible resource locally and then sends requests to the remote host, this should reduce the traffic to our CAS tier and improve the storage operations performance.

Reviewed By: krallin

Differential Revision: D36758399

fbshipit-source-id: 374804a3c9d5d736042088096e2d88557a12f337