Simple "definitely assigned" optimization
Summary:
When compiling:

```
def f(x): return x
```

Before it was compiled to:

```
load `x` to temporary
return temporary
```

Load to temporary in general case is needed to check that variable is assigned.

For parameters we know that it is definitely assigned, so load is not needed, and bytecode can be just:

```
return `x`
```

## To be done later

### Similar simple optimization can be done for `for` loop

E. g. when compiling loop body:

```
for x in ...:
  y = x # here we know `x` is definitely assigned
```

This is simple optimization, can be done later.

### Definitely assigned to regular local variables

```
x = ...
# here we know that `x` is definitely assigned,
# but there's currently no framework to do analysis like that
y = x
```

Also can be done later.

Reviewed By: ndmitchell

Differential Revision: D36765558

fbshipit-source-id: 1078ef13aaf8ee9b98c1a50b2c0c195c89d17922