Fix duplication of target deployment flags
Summary:
Deployment target flags for `apple_library()` should not be propagated above as linker flags - there should only really be a single deployment target when linking.

Instead, the deployment target for a library should only be used when actually linking the shared flavor of the library itself.

For example, imagine you have: `A (apple_binary) -> B (apple_library) -> C (apple_library)`. Before this diff, when linking `A`, we will end up with duplicating the target deployment flag 3 times.

For FBiOS, this is causing 21k occurrences of the same deployment flag which ends up breaking linking.

Reviewed By: chatura-atapattu

Differential Revision: D36805605

fbshipit-source-id: 01ab0eb399e76eb6bf1e5b5050399b5d860baae4