buck2: ignore artifact values in python manifest
Summary:
Building this right now is a 6.6GB action that basically only runs `echo`. It
takes forever as a result:

```
buck2 build fbcode//mode/opt fbcode//buck2/tests/e2e:test_build_isolated-manifest
```

The reason is because it uses `$(location)`, but only needs the path, not the
artifacts.

This diff updates this to be less bad by exposing a label that lets genrules
opt in to ignore artifacts...

Reviewed By: ndmitchell

Differential Revision: D36813040

fbshipit-source-id: 72572879f199e3ae5b8102c433eff2ab13af3ef5