buck2: executor: allow controlling whether we fallback on failure or not
Summary:
I'd like to use fallbacks if an action fails to execute on RE at all, but I
don't want to (always) try everything that fails twice.

This requires quite a bit of refactoring of the hybrid executor. The reason is
that currently the fallback logic in the hybrid executor assumes it can just
set claim rejected if it fails before making a claim (this happens on RE if we
fail before we get to the download for example), and then the select loop will
just wait on the other executor, but in the case where we only want to fallback
after a failure, this means that we end up returning with a claim rejected,
which is an error.

This code is a lot more complex than I'd like. The complexity stems from the
fact that the first call can fail with or without having made a claim, in which
case we want to either restart the secondary, or just wait for its call.

The problem is we don't know which happened (and we can't look at whether a
claim did happen as that would be racy), so we have to await the secondary and
then see if that yields a result or not, and if that results is a rejected
claim, then we have to retry.

FWIW, in full-hybrid, the current code didn't retry when the failure came from
the secondary first (it would just return that error). I think that's probably
not great. The new code should, I think, not have this issue (which is also why
it's more complicated, besides the desire to reuse code with the new fallback
path).

I'm not super happy with this. I think I want to update claims to allow pausing
execution of the futures instead of allowing the claim to only synchronously
say yes or no, which would make this *a lot* simpler, but for now, since we're
not actually using hybrid execution yet, this is a smaller refactor.

Reviewed By: ndmitchell

Differential Revision: D36808961

fbshipit-source-id: 436b36a7a12a69ff1b260a2853ea85cac2018284