buck2: remove CommandExecutorConfig from ExecutionPlatform's `Hash`
Summary:
I think it doesn't quite make sense to include in the hash of the command
executor config in the hash of an ExecutionPlatform, since this means that
e.g. changing metadata or whether the executor is hybrid we send to RE ends up
changing the hash.

FWIW, this looks like it's only used by target hashing: if I remove all the
`Hash` then this is the only thing that actually tries to hash a
ExecutionPlatform, transitively, so perhaps that should change to begin with?

Reviewed By: ndmitchell

Differential Revision: D36634563

fbshipit-source-id: 9548f01f73e2264888865fd076f605574eae3803