Check return type in function body
Summary:
Introduce new bytecode instruction `ReturnCheckType`.

When a function has return type specified, use this instruction instead of usual `Return` or `ReturnConst`.

This instruction checks the given value matches the function return type and then returns.

It should make code maybe insignificantly faster in common case (when there are no types), and a little slower when there are types (but that can be optimized later).

But now error location is reported properly.

Reviewed By: ndmitchell, bobyangyf

Differential Revision: D36786469

fbshipit-source-id: 48f54cd36f3ffefbd0b2b05e92890f3d0c873768