Use the configured target in xcode-data
Summary:
Previously, we were returning an unconfigured target in the `xcode-data` json. This meant later on, to do a lookup, we would have to create an unconfigured to configured map in order to lookup the right value.

Returning the configured target allows us to do a correct lookup and avoid the workarounds.

Reviewed By: johnhaitas

Differential Revision: D36828716

fbshipit-source-id: cfbe1e7551d944d08790a450f78d1df7ad6697a6