Add functionality to select which files to include in the project
Summary:
The crux of this stack.

We currently add files to the project and select which targets to prebuild in several places,  making the functionality hard to keep consistent. Here we add a method to select which files to include in the project, returning along with it, which targets to prebuild and and other files to include.

Specifically:
1. If a file is part of a target, and a non-generated file on disk, include it
2. if a file is generated, include it iff:
 a. the generating target meets the inclusion criteria (includes, transitively_included_targets)
 b. the generating target is not excluded (excludes and exclude_by_label)
3. if the file is generated, and the generating target included in the project (2 above) and is is a genrule, include any src files of the genrule

Point 3 specifically allows us to include remodel, cql files and other such files that are used by genrules to create output.

Reviewed By: johnhaitas

Differential Revision: D36828713

fbshipit-source-id: 433dd9b9274092b55477f2a2c571a1557915dd90