ExprLogicalBinOp
Summary:
Merge handling of `and` and `or` operations.

`and` and `or` are anti-symmetrical, for example, when compiling:

```
return a and b
return a or b
```

code is almost the same:

```
return b if a else a
return b if not a else a
```

Amount of code if reduced only a little, but uniform handling of and and or may reduce possible future error.

If you don't think this is improvement, I'm fine with not landing this diff.

Reviewed By: ndmitchell

Differential Revision: D36846923

fbshipit-source-id: 88c90ced3009a4e4576041f328dd558ab766d475