Implement instrumentation_test
Summary:
Per title - this is a basic implementation that aims to replicate how we execute instrumentation tests in buck1 (by constructing the same command line).

Right now, this will only work when running with Remote Execution - there is no support for local running with an emulator right now.

In a future diff I will read the test runner and package info from the manifest.

We also have to always do dynamic listing, because tpx has a hack to fallback to dynamic listing if static listing fails, and that hack does not work with buck2. Tracked in T122022107.

Reviewed By: krallin

Differential Revision: D36762544

fbshipit-source-id: 77e8fe8fff20583c998b7df57aee57090978be69