Make LSP jump to definition for loaded symbols
Summary:
For a given symbol, if a user clicks on "jump to definition", and that symbol is imported via the "load" statement, attempt to jump to the original definition in the loaded file, rather than the line of the original file that is just a call to "load". Fall back to that old behavior if an invalid filename is passed to "load", or the symbol is not found in the loaded file.

(Note: this ignores all push blocking failures!)

Reviewed By: ndmitchell

Differential Revision: D36425722

fbshipit-source-id: 2e7b10409a23b0bcb5fe056e35ed82d60ace564f