More tracking of definitely assigned locals
Summary:
Track variables which are definitely assigned, and when they are definitely assigned, use faster instructions for them:
* unchecked `Mov` instead of checked `LoadLocal`
* or no `LoadLocal` instruction at all if a local variable slot can be referenced directly by an instruction (e.g. when compiling `-x`)

This is continuation of D36850429 (https://github.com/facebookincubator/buck2/commit/0859f5ff46614aa8309d73dde07fc9088a10259d) which marked function parameters definitely assigned. This diff adds simple tracking of locals.

This diff implements relatively simple algorithm.

### 1. Function parameters

Function parameters are considered definitely initialized (as before)

### 2. Sequence of statements

If some variable is definitely accessed (read or written) in a statement, then it is definitely assigned in the following statement. For example, if there's a statement:

```
x[y] = f(z)
# here we assume than `x`, `y`, `f`, `z` are definitely assigned.
```

### 3. For loop

When entering loop body, we save the state of definitely assigned variables, and continuing compilation of loop body. When continuing compilation of code after the loop, we restore the state of definitely assigned variables which was the state before the loop. For example:

```
for x in y:
  # here we can assume `x` and `y` are definitely assigned
  z = 1
  # here `z` is definitely assigned as well
# here `y` is definitely assigned, but we don't know about `z`
```

### 4. If statement

This is similar to for loop.

Before compiling each of branches, then or else, we reset the state to the state before if, and compile the branch.

When compiling the statement following if statements, we reset the state which was before the if statement. For example:

```
if c:
  # here `c` is definitely assigned
  y = 1
  # here `y` is definitely assigned
else:
  # here `c` is definitely assigned
  y = 2
  # here `y` is definitely assigned

# here `c` is definitely assigned, but `y` is not, because we don't (yet) merge branches
```

### 5. Missing cases

Only the most common patterns are covered, some analysis are missing, for example:

```
(a, a) # we don't mark second `a` as definitely assigned
```

Or another example:

```
if cond:
  x = 1
else:
  return

# here we could know that `x` is definitely assigned
```

But these can be tweaked later.

## How safe it is?

If optimizations are done incorrectly there are two possibilities:
* program will crash on dereference of value
* program will report variable is not assigned in the wrong place

Memory corruption should not be possible, so it should not be too bad.

Reviewed By: ndmitchell

Differential Revision: D36874593

fbshipit-source-id: 3fac7f992b2a394ffb8004c2bda9df99bb774f64