Add config to enable "internal" javac in fbsource
Summary:
Enabling this requires the `-c buck2.enable_internal_javac=true`
and that there is no java.javac or tools.javac set. The buck2 flag will go away
when we are comfortable with the rollout and the other will stay (which matches
buck1 that will use internal javac unless you specify an external one).

Reviewed By: IanChilds, mykola-semko

Differential Revision: D36438618

fbshipit-source-id: 63a7f99a4f2e2fbb72c47839f64d35d3dd6c5152