Add framework for enabling "internal" javac
Summary:
This adds a "javac_protocol" to the toolchain and the logic to enable
a javacd path for compile_to_jar (that currently just fails). This will be used
when a toolchain doesn't have a javac set (which our current config_backed toolchains
would never hit because they set a non-None default for javac, but the next diff
changes that).

Reviewed By: IanChilds, mykola-semko

Differential Revision: D36438616

fbshipit-source-id: 196bea6e8c161ea2529475d96a1a292de15e3ca0