Fix quadratic performance of queries with '%s'
Summary:
targetset `.union()` creates a new set that's the union of the two.
For a long chain of these, this has bad performance. Instead, union them
in-place.

We should likely improve this in the queries themselves as we know that people
construct long chains of `x + y + z + ...` (for example, we know that we get
stack overflows if we just recurse for such expressions). This will take a
bit of work as it'll probably mean making query evaluation support functions w/
mutable params.

Reviewed By: stepancheg

Differential Revision: D36959228

fbshipit-source-id: 9f9843d5a3a1b749ca70a17158ff5be01f48d1be