loading package -> loading package file tree
Summary:
"loading package" is misleading. What is package? One may think that package is everything in the package: files + BUCK evaluation + targets construction.

But "loading package" event is only about loading file tree.

Which means I/O, not starlark or something else in Buck.

So it might be helpful to rename it to be more clear about what is going on inside Buck.

Reviewed By: krallin

Differential Revision: D36961824

fbshipit-source-id: 465ce1e9d3b35a214602f56bfb179b9e3b7c41f2