Basic android_instrumentation_apk implementation
Summary:
`android_instrumentation_apk` is an APK that can be used as the "test APK" in Android instrumentation tests (https://developer.android.com/training/testing/instrumented-tests).

It is similar to a normal Android APK, except that any resources, native libraries and Java code that is already present in the APK-under-test that it builds against are stripped out of the APK.

For now, we just build an APK that contains all resources, native libs and Java code, and we will work on stripping them in T122203218.

To use this in `android_instrumentation_test`, add a provider `AndroidInstrumentationApkInfo` which holds the `apk_under_test`, and use that if it is present - if not present, then there is no `apk_under_test`, it is just the same as the test APK.

Reviewed By: cjhopman

Differential Revision: D36779002

fbshipit-source-id: 8f846e8cab35344bd63e878100f34aa0e0d2fa2c