Refactoring BuckTargetHash from an enum to a u128
Summary: In the code for Buck2's target hashing, there is a BuckTargetHash enum which is Buck2's representation of a target's target hash. This enum has 2 values, compatible and incompatible. However, the code has been refactored and now all targets returned by get_compatible_targets are always compatible, so this enum is no longer needed - BuckTargetHash only needs to contain a u128, the type of the hash.

Reviewed By: stepancheg

Differential Revision: D36911338

fbshipit-source-id: 0b7811b262f3b85d6226f2cc7cd19f3ea1c6f26a